+++ Logic/Logic.vcxproj
@@ -286,6 +286,7 @@
     <ClInclude Include="include\Player\Skill\SkillShieldCharge.h" />
     <ClInclude Include="include\Player\Weapon\ModelAnimation.h" />
     <ClInclude Include="include\Player\Weapon\Weapon.h" />
+    <ClInclude Include="include\Player\Weapon\WeaponAnimation.h" />
     <ClInclude Include="include\Player\Weapon\WeaponFreezeGun.h" />
     <ClInclude Include="include\Player\Weapon\WeaponManager.h" />
     <ClInclude Include="include\Player\Weapon\WeaponMeleeSwing.h" />
@@ -389,6 +390,7 @@
     <ClCompile Include="source\Player\Skill\SkillManager.cpp" />
     <ClCompile Include="source\Player\Skill\SkillShieldCharge.cpp" />
     <ClCompile Include="source\Player\Weapon\ModelAnimation.cpp" />
+    <ClCompile Include="source\Player\Weapon\WeaponAnimation.cpp" />
     <ClCompile Include="source\Player\Weapon\WeaponFreezeGun.cpp" />
     <ClCompile Include="source\Player\Weapon\WeaponMeleeSwing.cpp" />
     <ClCompile Include="source\Projectile\Projectile.cpp" />
